restructuring of directories and files
