+++ src/engines/auth_flood.rs
@@ -1,6 +1,6 @@
 use std::{thread, time::Duration};
 use crate::iface::{IfaceInfo, InterfaceManager};
-use crate::pkt_kit::PacketSniffer;
+use crate::sniffer::PacketSniffer;
 
 
 
@@ -13,7 +13,12 @@ impl AuthenticationFlooder {
     pub fn execute() {
         let iface = "wlp2s0".to_string();
         InterfaceManager::enable_monitor_mode(&iface);
-        println!("iface down");
+        
+        let mut sniffer = PacketSniffer::new(iface.clone(), Self::get_bpf_filter());
+        
+        sniffer.start();
+        thread::sleep(Duration::from_secs(3));
+        sniffer.stop();
     }
 
 
@@ -21,7 +26,7 @@ impl AuthenticationFlooder {
 
 
     fn get_bpf_filter() -> String {
-        "ether[0] & 1 = 1 and ether[1] = 0x50".into()
+        "type mgt and subtype beacon".into()
     }
 
 }
